Lab 4 exercise of SLACK integration with ServiceNow
The goal of lab 4.0 is to create and close Slack group using REST Messages so when a priority 1 is submitted a new channel is group is created and when this Incident is resolved this group will be archived. New field will be created on the Incident table to store the group’s Slack identifier, this ID will be received from the payload response when the REST Message is sent.